+++ src/main/kotlin/sound/frequencyBins/FrequencyBin.kt
@@ -1,3 +1,3 @@
 package sound.frequencyBins
 
-data class FrequencyBin(val frequency: Float, val amplitude: Double)
+data class FrequencyBin(val frequency: Float, val amplitude: Float)
